Обновлен ffmpeg n8.1-dev-629-g8a34faa250. Файл libavcodec\vp9.c обновлен частично.

+++ src/ExtLib/ffmpeg/libavcodec/hpel_template.c
@@ -29,9 +29,9 @@
 static inline void FUNC(OPNAME ## _pixels8_l2)(uint8_t *dst,            \
                                                const uint8_t *src1,     \
                                                const uint8_t *src2,     \
-                                               int dst_stride,          \
-                                               int src_stride1,         \
-                                               int src_stride2,         \
+                                               ptrdiff_t dst_stride,    \
+                                               ptrdiff_t src_stride1,   \
+                                               ptrdiff_t src_stride2,   \
                                                int h)                   \
 {                                                                       \
     int i;                                                              \
@@ -40,9 +40,9 @@ static inline void FUNC(OPNAME ## _pixels8_l2)(uint8_t *dst,            \
         a = AV_RN4P(&src1[i * src_stride1]);                            \
         b = AV_RN4P(&src2[i * src_stride2]);                            \
         OP(*((pixel4 *) &dst[i * dst_stride]), rnd_avg_pixel4(a, b));   \
-        a = AV_RN4P(&src1[i * src_stride1 + 4 * sizeof(pixel)]);        \
-        b = AV_RN4P(&src2[i * src_stride2 + 4 * sizeof(pixel)]);        \
-        OP(*((pixel4 *) &dst[i * dst_stride + 4 * sizeof(pixel)]),      \
+        a = AV_RN4P(&src1[i * src_stride1 + 4 * (int)sizeof(pixel)]);   \
+        b = AV_RN4P(&src2[i * src_stride2 + 4 * (int)sizeof(pixel)]);   \
+        OP(*((pixel4 *) &dst[i * dst_stride + 4 * (int)sizeof(pixel)]), \
            rnd_avg_pixel4(a, b));                                       \
     }                                                                   \
 }                                                                       \
@@ -50,9 +50,9 @@ static inline void FUNC(OPNAME ## _pixels8_l2)(uint8_t *dst,            \
 static inline void FUNC(OPNAME ## _pixels4_l2)(uint8_t *dst,            \
                                                const uint8_t *src1,     \
                                                const uint8_t *src2,     \
-                                               int dst_stride,          \
-                                               int src_stride1,         \
-                                               int src_stride2,         \
+                                               ptrdiff_t dst_stride,    \
+                                               ptrdiff_t src_stride1,   \
+                                               ptrdiff_t src_stride2,   \
                                                int h)                   \
 {                                                                       \
     int i;                                                              \
@@ -67,9 +67,9 @@ static inline void FUNC(OPNAME ## _pixels4_l2)(uint8_t *dst,            \
 static inline void FUNC(OPNAME ## _pixels2_l2)(uint8_t *dst,            \
                                                const uint8_t *src1,     \
                                                const uint8_t *src2,     \
-                                               int dst_stride,          \
-                                               int src_stride1,         \
-                                               int src_stride2,         \
+                                               ptrdiff_t dst_stride,    \
+                                               ptrdiff_t src_stride1,   \
+                                               ptrdiff_t src_stride2,   \
                                                int h)                   \
 {                                                                       \
     int i;                                                              \
@@ -84,9 +84,9 @@ static inline void FUNC(OPNAME ## _pixels2_l2)(uint8_t *dst,            \
 static inline void FUNC(OPNAME ## _pixels16_l2)(uint8_t *dst,           \
                                                 const uint8_t *src1,    \
                                                 const uint8_t *src2,    \
-                                                int dst_stride,         \
-                                                int src_stride1,        \
-                                                int src_stride2,        \
+                                                ptrdiff_t dst_stride,   \
+                                                ptrdiff_t src_stride1,  \
+                                                ptrdiff_t src_stride2,  \
                                                 int h)                  \
 {                                                                       \
     FUNC(OPNAME ## _pixels8_l2)(dst, src1, src2, dst_stride,            \
